Main changes from version 0.14:
- Tweaked transit timings a bit more (lengthen pre-T.4 silence by 10 ms)
(see PAN-00754 and PAN-00667 for a detailed discussion).

- Changed to the NSF codes allocated by the DTI's representative,
Alan PUGH (tel. 0118 9210909): country code 0xb4, country subcode 0x00,
manufacturer code 0xb6, where the convention used is that the bits
are transmitted starting with the least significant bit of the octets
as given.  Note this means prior versions of !Fax will not detect or
be detected by this version of !Fax.

- Changed the page chop overflow tolerance to 3% (see PAN-01364 for
a detailed discussion).

- Allowed DCS to be followed by DIS when transmitting when the DCS was
caused by RTP, RTN or EOMthenMCF (see PAN-01686 for a detailed
discussion), such that the receiver's capabilities are allowed to
change during a transmission.

- Allowed clients such as !Daytona to request a halving of the datarate,
when receiving, where possible; this allows clients to try to stem
an incoming datarate they can't cope with because e.g. the outgoing
datarate is lower (see PAN-01686).

Tagged as version 0.15.