* These changes make no difference to the build of the TextGadget module.
Re-arranged header files to be suitable for sharing with Toolboxlib and
for export to RiscOS/Export/Lib/tboxlibs/h.  The ordering and style of
these files now matches other toolboxlib header files.

Private TextArea macro definitions moved from TextArea.h to TextAreaP.h

Most function declarations are now auto-generated by MethodGen so they
come with documentation for free, except for TextArea_GetSelection*
which need to be done manually.  This new set of declarations is
identical to the hand-written ones which existed previously.