This log is in three parts, this being part 2, because CVS falls over with big logs - the rest of the files in this release are checked in with a message similar to the following:
> Check-in of Browse v2.06; using very small log file to try and avoid
> CVS crashing. Some of the Resources may well be out of date due to CVS
> locks being in place after earlier server-end core dumps, which I can't
> remove myself.
>
> I'll try and check one file in at the end with the full change log so
> people know what's happened (it's reached about 16K...) - I'll make
> that the TaskObey file '!All' in the top level directory. So for the
> full log, look for the changes on this file.

More information on what this release is, and how to build it, can
be found in part 1 of the log.

 * JavaScript                      (Part 1)
 * Resources                       (Part 1)
 * File I/O, data format support   (here)
 * Reformat and redraw             (here)
 * Fetching                        (here)
 * Image system                    (Part 3)
 * Plug-ins                        (Part 3)
 * Build environment               (Part 3)
 * Miscellaneous                   (Part 3)


File I/O, data format support
=============================

Text export now has optional "with or without embelishments", the default
being for just a plain text file (no references at the bottom, no table
structure information).

You can drag text files to form writables (single and multiple line,
but NOT password items. It's trivial to add the last bit but I
think it'd be a bad idea. Data is appended up to the size limit
specified by the field (unlimited for multiple line items). Holding
down Ctrl replaces the current value entirely (still limited by maximum
ield length). The caret is moved from wherever it currently is and
placed at the end of the added data. Note that you *cannot* insert
in the middle of text by placing the caret and dropping.

Things can get distinctly sluggish when you start dragging many Ks of
text file into a multi-line writable, so the keyboard buffer is cleared
for any input code < 32, equal to 127 (backspace - this one is a little
less clear cut, but I think it's probably worthwhile) and anything
> 255. This way, when typing, you don't lose characters; but when moving
around, you don't find the cursor "running away" from you.

Single character leafnames (e.g. "a.class") used to be reported just as
"Index" in a save dialogue for the object, but are now reported correctly.

The fetcher, when saving data to a file, gets the filetype from the server.
If this is FileType_DATA or 0x000, it sees if the MimeMap module can get
anything better from the filename extension (if present). This is now done
for returned types of FileType_TEXT, too. So, when saving an inline image
which is a Draw file out in its original format, you get a Draw file, not
a text file... Provided you have this in your MimeMap file:

image/x-DrawFile	DrawFile	aff	.draw	.drw


Reformat and redraw
===================

Yikes - ever since markers, "reformat page to fit paper" hasn't done
anything when printing; the reformatter reset the HTML page limits to
the window size at startup! Fixed. Using markers, you can now also reformat
from the top of the visible area and still reformat to fit the paper; the
print dialogue used to grey out the reformat option in this case. It still
does this when printing to the bottom of the visible area, though.

When following a link, if RefoKeep was 'yes', the page could sometimes
jump to a pre-stored scroll position; it was down to the resize routines
remembering it, but if windows_check_reformat() or windows_remember_size()
didn't actually do a reformat the position was not forgotten. It is now,
so this jumping should no longer occur.

Vertical scroll bars no longer get stuck in frames once they've arrived -
the routines that start a reformat due to window resize etc. set a new
flag, "tools_lock", in the browser_data structure which tells the tool
check routine to allow vertical bar removal (it then clears the flag to
stop recursive reformatting problems). It still isn't 100% reliable -
the vertical bar can appear when it shouldn't and then go on the next
reformat or reload - but it's definitely a lot better than it was.

HSPACE and VSPACE are implemented for IMG and OBJECT elements.

Horizontal gaps in table cells solved; reformatter was ensuring any line
is a multiple of 4 OS units high, to ensure correct pixel alignment,
rather than using wimpt_dy().

Tiled images were being plotted in the wrong place in some modes (a '+4'
in the starting y coordinate calculations in the three (!) filler / tile
functions should have been '+wimpt_dy()').

Image placeholders used to be 3D unless <= 8x8 OS units in size, in
which case they are a 1 pixel wide 2D black outline; they now have to
be greater than 8 wide and 30 high. The graphics rectangle used for
ALT text within this placeholder is larger for the 2D case as well,
now. This means you can typically see more ALT text on small images.

Multi-line writables (text areas) would redraw a vertical column taller,
or horizontal column wider than the element covers itself should the
user scroll enough in any one direction (e.g. Ctrl+Up or Ctrl+Down to
go to the top or end of the area). The code working out how much of
the area to redraw by the number of lines or columns moved didn't ensure
it kept within the area. Because an update rather than force redraw is
done, with background blending turned on text around the area would
"thicken". Now fixed.

<P> elements could fail to work as expected in some circumstances. For
example, there would be no gap above '<p> <b>...' - the <b> text doesn't
have the <p> attached, the tag before it does. However, the white space
on this tag gets stripped so it ends up with a NULL text field. This
means that paragraph space is allocated but the text height for the line
thus far is zero. This has been fixed through special case code in
reformat_check_height.


Fetching
========

Proxying - you can now have Choices file items "UseProxyxxx" and
"ProxyAddressxxx" where 'xxx' is the protocol name to proxy, in lower
case - e.g. "UseProxygopher" and "ProxyAddressgopher". HTTP proxying
is done through just "UseProxy" and "ProxyAddress" for backwards
compatibility, with the Choices dialogue options affecting only
HTTP proxies. In addition, "ProxyExclusions" contains a comma separated
list of domains to exclude from any proxying where each item starts
with a full stop, or a specific host to exclude if not - e.g.
"www,.ac.uk" would proxy anything set up in Choices with the exception
of a host called 'www' (e.g. "http://www/browser/") or any host name
ending in '.ac.uk' (e.g. "http://src.doc.ic.ac.uk/"). There is no front
end control for anything other than the HTTP proxying as in earlier
versions of browse; being unable to find a relevant entry in the Choices
file for proxy settings is taken to mean "don't proxy this protocol".

browser_base_url() function added, which through browser_base_callback
can supply a base URL for the current document.

In-page references (URLs with '#something' in them) would fail if the
page contained a BASE tag which pointed elsewhere (that is, the page
would be fetched from its base URL and references would work thereafter;
not very useful for developing sites on your local hard drive). Fixed.
Fixed also the problem where in-page references would fail once after
a forced reload.

Submitting a GET form back to the same URL would lead to the query
string getting appended to the fetch URL over and over. form_build_data
now checks to see if there's already a query string in the submission
URL for GET methods and if so, strips it. fetchpage_new_add will do a
similar thing in the case of the additional data starting with '?'.