A couple of superficial changes to the file structure have resulted in CVS needing this complete deletion/recreation.
Basically, the IOCtl code is getting too big to be tacked onto the
DeviceCall file. So this has been split into the file 'IOCtls'. Given that
this alone is enough to stuff CVS, the opportunity was taken to rename the
'Init' file to 'GetAll', as it doesn't init anything, and Gets all the files
, GetAll being the traditional name for such a file.

Additionally, a Docs directory has been added, which contains a copy of the
functional spec, which will hopefully be kept up to date. Also this directory
is for miscellaneous files which may aid understanding of the thought processes
which led to various bits of code or design decisions, namely the Ursula file
at the present moment.