Internet 5.08	Fixed memory leak in routing table.
net/route.c changed in RTM_ADD case, line 491.  Else clause added to free
allocated rtentry structure.  Fix pulled in from FreeBSD 1.37 version of
net/route.c.  This is the only fix taken from the 1.37 version.  1.37
version used as it is the one used in the latest FreeBSD, 2.2.6.

This is to fix a memory leak in Daytona on PPP dialups.