The scope for this work is described in 2205,206/DS, presently visible at the following URL:
http://www.internal/~babel/funai3/specifications/software/osk-funai3-ds.html


New features
============

Resources for Japanese layout added. These make full use of the extensions
detailed below. There's also a new keyboard layout in Resources created
explicitly for testing / demonstration purposes.

The Makefile has been left in a state where it defines FUNAI3, i.e.
creates a "new style" module. This should be 100% backwards compatible
with old style resources.

Supports a general number of layers from 1 to 'n' limited by available ROM
or RAM for sprites and the Messages file, and RAM for an array to hold the
default gadget numbers for each layer. Sprites for the layers may vary in
size; the window will resize to cope, preserving bottom y coordinate for
-scrollup and top y coordinate for -scrolldown. BOUND_MOUSE builds will
have the mouse reconstrained. If the pointer falls out of the gadget used
to change layer because of the change in size it is returned to the middle
of that gadget. The geometry change message is rebroadcast if the window
size alters.

For designs with other than 2 layers, a Messages file entry "Layers" is
read to find out how many there are (e.g. "Layers:4" to enable layer
numbers 0 through 3 inclusive). The minimum value is 1.

Key definitions in the Messages file can be of the form "KeyLxxx" where
L is a layer number, rather than "KeyQxxx" or "KeyAxxx" - see the section
on deprecated features for more. They can contain strings if preceeded by
one double quote character (e.g. 'Key2004:"www.' would make component 4,
layer 2 output the string "www."). This can include characters with codes
128 to 255, thus allowing UTF-8 byte streams to be specified directly.


Bug fixes
=========

If any the Messages file entries "GadgetNoA", "GadgetNoQ", "GadgetNoEB",
"GadgetNoCL" or "GadgetNoSh" were missing, a value would be assigned for
that function equal to whatever happened to be in the messages lookup
buffer at the time (e.g. a build without Caps Lock could have accidentally
assigned this to the default gadget number for the QWERTY keyboard, and a
build without an extension toolbar could have accidentally deleted a
gadget).

If oskb_click_handler() encountered an unknown layout it would report the
error but then try and process the click anyway, which would still either
fail or at best behave unpredictably. Now returns 0 from the handler
function after reporting the error.

A "sort-of" bug in IconHigh has been worked around. The "ignore next
character" SWI is documented to ignore the next *character*, but in fact
it just ignores the next *byte*. This is a problem when you're outputting
a zero byte to escape a subsequent byte in the range 128 to 255; you must
call Ignore for the zero byte, and Ignore again for the second byte.
IconHigh doesn't realise that a zero byte signals another is due to make
a whole character.

Various places assumed a null component ID is 0. It is actually -1, or
better still, use the #defined value of NULL_ComponentID. For example,
Toolbox_ShowObject was being called without the #defined values for the
show type and with 0 for the parent object and component. This has been
corrected to NULL_ObjectId and NULL_ComponentId.


Deprecated functions
====================

Special key names used just to output strings (e.g. "www", "com") are no
longer needed (see 'New features' above) though they will still work.

Messages file entries "SpriteL" where "L" is the layer number should be
used to record the name of the main display sprites used for each
layer, though "kb_alpha" will be used for layer 1 and "kb_qwerty" used
for all other layers if these are absent (thus mimicing previous
behaviour).

Key definitions in the Messages file used to be "KeyAxxx" or "KeyQxxx".
They now they take the form "KeyLxxx" where L is the layer number from 0
upwards. The gadget code "xxx" is still a three digit decimal number
regardless of how many digits the 'L' component of the token name
occupies. Key0xxx and Key1xxx will be tried, and if not found, KeyQxxx
or KeyAxxx are tried instead.


Removed functions
=================

None strictly. However if FUNAI3 is defined when NCKeyboard is built, the
extension toolbar is still supported, but will only function correctly if
all layers are the same size. You can't have a toolbar and varying height
layers at the moment. It could be done by giving a range of component IDs
over to the toolbar and shuffling them vertically, but this is beyond the
scope of the work in 2205,206/DS.


Misc
====

Copyright messages in all sources changed to "Acorn Computers Ltd." from
"Acorn UK Ltd.," (sic). "Relevant documents" sections mention later specs,
in a general fashion. History entries for these changes refer to CVS logs.
All globals now declared in NCKeybd.h; no need to have "extern xxx" all over
the place in the .c files. All cases of "exit(0)" or "exit(1)" changed to
"exit(EXIT_SUCCESS)" or "exit(EXIT_FAILURE)" respectively.

Extension to oskb_calc_sprite_geom (private function in windowops.c) - it
takes a layout (OSKB_ALPHA / OSKB_QWERTY for non-FUNAI3 builds, else a
layer number) from which it works out which sprite to measure. This gives
no behavioural difference for non-FUNAI3 builds but is essential otherwise
to work out the tallest sprite present as well as the starting sprite.
To help support this and because it is useful elsewhere, there's a new
function to get the sprite name from a "SpriteL" form Messages token or
a default value in their absence present: oskb_get_sprite_name, oskbutils.c.

Extension to oskb_change_layout - it takes a gadget ID or NULL_ComponentId
which is used for the 'reposition the mouse pointer if it falls out of the
gadget' function.

For maintaining top y coord when -scrolldown is specified, scroll_point
is now a global rather than local to main.c.

A lot of NC2 resources !Run files changed. They were requiring KeyWatch but
not IconHigh - when the application requires IconHigh but not KeyWatch!


Admin
=====

I've tested using the NC2 Denmark resources and the Test resources as
examples of old-style and new-style resources respectively. Building with
DFLAGS clear works with the Denmark resources, fails (by having the bitmap
displaced and no key mappings, but doesn't crash) on the new resources.
Building with DFLAGS = -DFUNAI3 works fine with both resource sets. With
DFLAGS = -DBOUND_MOUSE it works as DFLAGS empty but bounds the mouse as
expected, and with -DFUNAI3 and BOUND_MOUSE it works again as expected
with both resource sets working, showing correct mouse bounding behaviour.

For reference (with CC 5.07, link 5.07):

  DFLAGS                   RAM load module binary size, bytes

  Empty                    15004
  -DFUNAI3                 16400
  -DBOUND_MOUSE            15232
  -DFUNAI3 -DBOUND_MOUSE   16736