ErrorLib 0.07
Essentially a new library, but remains backwards compatible with ErrorLib
as on the Daytona branch.  If -DERRORLIB is included then errors returned
by SWIs and Toolbox calls can be output with filename and line number to
supplement the original error.  The output is sent to standard Error boxes
or Throwback windows.  If -DERRORLIB is not defined then the original error
is output to an Error box in the same fashion as previous versions.