Added the Norwegian resources, including all HTML and graphics.
Changed the introduction screens to include the DORO "geezer-in-a-seat"
graphics: this change has been applied to all the resource sets.

Added one additional screen to each resource set: this is used to
request that the original smartcard is inserted into the unit on
dial-ups when the smartcard has been exchanged.

Version 0.19. Tagged as 'Funai2UI-0_19'