EmailLib 0.44
Major type name changes to remove NetFax specifics from library.  All NFE
prefixes are changed to ELIB.

BOOL, TRUE and FALSE changed to correct c++ bool, true and false.

Shared/Common Email library types moved to a new library in:
RiscOS/Sources/Lib/EMail/Common

EmailLib is now dependant on this new library.

This version of EmailLib does not work with any current version of NFELib
(NetFax Email library - version at this time is 0.49).  NFELib needs to be
modified to use ELIB type names internally in it's calls to EmailLib.  The
last of the NFE code in EmailLib, the X-NetFax header types, will be moved
to NFELib once NFELib is updated and the EmailLib development branch changes
are merged in to EmailLib on the TRUNK.

Current NFELib can continue to use the EmailLib on that's on the Daytona
branch with it.