Updated the "!Run" files in each set of resources to include aliased command "NCFrescoRedial".
Incorporated the translation for "Service Provider" into the dialling,
connecting and verifying screens in the Swedish resources.  This change
will also be applied to the other resource sets once the translations
become available.

Updated the UK NVRAM tags: changed the country setting to specify "UK".

Version 0.27. Tagged as 'Funai2UI-0_27'