EmailLib 0.49
New features
------------
Create reply functionality added - see rfc822msg class, you can for a
rfc822msg object create a new skeleton rfc822msg object with the To and
subject headers setup for you and the first text entity quoted ">" with
the .sig removed.  Reply and reply to all is supported

from header class changed to support the create reply functionality so that
to headers can be created from a from or reply-to header object of any type
(memory or file) and subsequently have mailbox objects added to and removed.

BFS (Breadth First Search) algorithm added to bodymulti/MIMEmultip classes for
searching for the 1st text entity in a message at the highest level for use in
replying.

Bug fixes
---------
WriteOut() to memory version fixed on all classes so that the 'walk' down
all the objects on WriteOut() appends to the memory and doesn't keep
writing from the start as it used to.  Some possible faults still remain
in here though.

Fixed memory leak in rfc822msg objects of the mime_obj in a message not being
destroyed.

Fixed WriteOut on MIMEentity so that a WriteOut doesn't permanently change the
stored Content-Transfer-Encoding of a MIMEentity due to the WriteOut which may
need to encode the object for the WriteOut operation.

Tests
-----
Changes tested with EmailLib test harnesses and in !NCMail.  Test harnesses
all pass, however !NCMail exhibits a malloc heap corruption problem.
Probably EmailLib fault, but not proven yet.

Misc
----
See !ChangeLog, .h and .c++ files for full details of the changes.