EmailLib 0.50
Actual MemCheck and HeirProf targets added now - it doesn't use the debug
target any more.  Therfore, you don't have to do a clean and swap between
builds.
    <export$dir>.Lib.Email.EmailLib.o.EmailLibMC is MemCheck build
    <export$dir>.Lib.Email.EmailLib.o.EmailLibHP is HierProf build
  HierProf build is just a normal build, but with function names embedded.

mailbox parsing fixed so that local (ie. addresses with no domain) email
addresses are parsed correctly.  mailbox GetData (used by GetAddress and
GetRealName) fixed for version when mailbox object is requested to allocate
the memory for the Get.  The output parameter, wasn't!  It's now correctly a
reference to a char* in GetData().  GetAddress and GetRealName were already
correct.

rfc822msg::CreateReply extended to take an optional parameter: exclude_address
which is an email address to exclude from the reply address.  eg. to allow
creation of a reply-to-all but the recipient of the original message.