AcornSMTP 0.32
Service_WimpCloseDown support added.  If the client passes its taskhandle on
Register, then if the client subsequently dies without calling DeRegister
first, the module can clean up.  Only works for clients which register for
WimpMessage notification.  Support needs to be added to the module in the
future to do this for Wimp apps. that don't request WimpMessage notification.

Tested with testharness and Mailbox library test harness.