Version:	0.33 Date:		1998-11-23 Name:		Richard D. Warren CVS tag:	rwarren_AcornPOP3_Client-0_33
Flag added to GetMessage and GetXLinesOfMessage, bit 1.
AcornPOP3_NO_SPACE_CHECKING defined in AcornPOP3.h for it.
If this bit is set it means that for retrival to file (not supported for
memory) the module will not use the filespace size value passed to it in
R3, but will write the data to the file until the end is reached or a file
system error occurs (ie. out of space).

Module initialisation fixed to initialise C++ runtime by calling ______main()
from module initialisation routine.

Tested with test harnesses and with Mailbox library 0.21 in !NCMail.

s: ----------------------------------------------------------------------