Changes:
	Made the exitoff command for the email client point to the
emailselectusernak transition, which will take the user back to the 'connect'
screen and state.

	The 'pleasewait' screen is now only displayed on non-ethernet builds.
It was getting displayed on an ethernet build and preventing the user from
doing anything else.

	Altered a couple of routines in ncma_scard.h which were using a
variable named 'new', which throws up a warning under the new C compiler.
Changed the variable to 'new_pin'.

Version 0.10. Tagged as 'NCMAlite-0_10'