Changes:
 	Made it so that when the smartcard is pulled out whilst in offline
email mode a dial request is delayed until the user inserts the correct card.
This is the same code as happens with a redial request whilst in
state_booted, so I have abstracted it into another routine, start_redial().

 	Altered the 'pleasewait' screen so that it is only displayed if the
bnoscreenchanges flag is unset. This prevents a dial request from within the
email client blanking the screen and confusing the email client UI.

 	Adapted the flags passed to the routine fetch_sc_bootblock() so it is
more apparent whether it was called with regard to the ISP_Capture module or
the email client. This whole thing is a little messy, and the NCMA Swi
MiscOP should really be tidied up sometime...


Version 0.12. Tagged as 'NCMAlite-0_12'