This change fixes fault report KJB981117a where the wait for dialtone option on the UI was being completely ignored.  The fix was verified by looking at the cvs log differences between version 0.98 and 0.97 on the trunk, and looking at the corresponding differences in the code (between versions 0.98 and 0.97 on the trunk).  Also verified with JBeranek, and tested by soft loading from zip drive by looking at the diallerb log on the command line to see if ATX3 had been inserted.
It is not necessary to merge this branch back onto the trunk because
this fault has already been corrected on the trunk, and there are no
other differences (other than adding the latest release notes).

Because this is a final release with only half a days testing at the
most before 10,000 mask ROMs are blown, we want to keep changes to a
minimum.  Hence the lack of inclusion of other changes that were made
on the trunk since NCUpgrade-0_85.

(The "no other differences" mention above refers to changes made in
the branch relative to the trunk version 0_85, i.e. the only change
in this deviation from the trunk is to correct the wait for dialtone
fault, which has been fixed on the trunk version).

Version 0.85, 1.83.2.1. Tagged as 'NCUpgrade-0_85-1_83_2_1'