Added another call to ncma_network_init() so that it gets called before going into the offline email client. This fixes bug RDW-017.
 	Added some basic debugging to some routines in ncma_utils.c to allow
tracking of the hourglass being turned on and off. Added appropriate
hourglass_on() and hourglass_off() calls to allow the hourglass to be
displayed once connecting. This is work in progress to fixing bug ATFUN-013

 	Fixed another little bug(let) in the DNS code on advice from RDW - if
there is not a backup DNS on the smartcard it produces a nonsensical DNS as
the backup DNS. Now altered so if there is no backup DNS it passes on only
the primary DNS.



Version 0.14. Tagged as 'NCMAlite-0_14'