Changed the exitoff string passed to the email client so that the ncma:emailselectusernak command is lowrcase, due to strstr() being case-sensitive.
 	Bracketed the ncma_dial() command with the hourglass. This ensures
that the hourglass appears whilst the NCDialUI_Start SWI is called, which is
by far the longest pause in the whole process. This is a definitive bugfix
for ATFUN-013 and ATBrowser-031.

 	Removed the code added in Version 0.14 which displayed the hourglass
whilst building the smartcard cache - this has been superseded by the
alteraion above.

 	A bug where redialling in the browser without the smartcard having
been removed did not fetch the new page has been fixed - the bRedialRequested
flag in the new start_redial() was not being set.

Version 0.16. Tagged as 'NCMAlite-0_16'