Forgot to mention at the last check-in the "idiot detection code" (ahem) in main.c, so modified a comment at the top to get another CVS log going. This checks the caps lock, shift key and extension bar sprite holder gadget numbers in the Messages file - if they're zero, it assumes the author meant "no gadget" (should've written '-1', of course).
If you really do want to use zero as an ID for one of these items, then put
the token "AllowZeroIDs" in the Messages file - the value is irrelevant -
which disables the above code.

Previously the lack of this code would mean that some builds showed no
keyboard sprite on old, legacy Messages files as they told NCKeyboard to
use gadget 0 for one or all of the caps lock, shift key and extension bar
sprites... Nasty. Note that this code is compiled whether or not FUNAI3
is defined at build time.