Fixed bug RDWNCMA-026. Created a new routine, select_screen_to_move_to(), which on a logout or exit from the email client decides whether to go to the connect or welcome screens. This is an existing puece of code that has been moved out to its own seperate routine so that it can be called from many places.
 	Rationalised the dialogue box handling. Now there are two routines,
ncma_dbox_open_to_target() and ncma_dbox_close_target() which replace
ncma_dbox_open() and ncma_dbox_close() respectively and have the target to be
opened or closed passed in.
	This alteration allows much more flexibility in the future and also
means that we can remove much duplicated code - for instance
ncma_standby_dbox_open() and ncma_standby_dbox_close() are not needed anymore.
	As part of this change, went through all the code and change routine
calls appropriately.

	Admin
	=====
	Have thoroughly tested all these changes on a Funai2 3-07 based ROM
image, both as part of the image and as a softload.

Version 0.25. Tagged as 'NCMAlite-0_25'