Trivial code change to make non-FUNAI3 builds behave the same way as FUNAI3 builds; don't throw an error and exit if asked to run with an extension toolbar but the kb_extension sprite doesn't exist, just carry on without the toolbar. Requested for STB3.
Removed - again - redundant o, oz, rm, abs, and aof directories,
with the '.cvstag' hack files inside them. The build system of this
software was modified some time ago to build properly and not need
such cumbersome work-arounds; these directories are created on the
fly when building from clean. It is safe (and indeed recommended) to
use MkClean before checking things back in.

Version 0.43. Tagged as 'NCKeyboard-0_43'